Amended subscripts on matrix elements
I think that the time subscripts in the X and A matrices should now be in terms of numbers rather than relative to 't' when displayed in this format. This is because the Y matrix is also specified in terms of time indices so the X and A matrices will need to match. I updated the subscripts to follow this notation.